Show errors when '+=' and '-=' operators are badly used on local variables
Make validators check usage of those operators on local variables to show
errors when they are called on 'self' or inexisting variables.

It was too time-consuming to implement a proper type-checking now, but
a warning is now shown to the user to warn him that ALE cannot guarantee
type safety, e.g. in the following statements:

    Integer a := 0;
    a =+ 5;
    a -= 12;