Set "*.xmi" as the default file pattern when running ALE
When ALE is run from the context menu it first opens a dialog
asking for a model instance to execute. This dialog now has
"*.xmi" for initial pattern so that XMI files are shown by
default without requiring any user typing.

An attempt was made to automatically select the XMI file that
is next to the DSL file by default (if any), but a bug seems
to prevent this feature, see:
 - https://bugs.eclipse.org/bugs/show_bug.cgi?id=214491