Homogenize "Hello World" projects (#71)
* Force Git to keep empty src/ folders

Some src/ folders are empty. To force Git to keep them anyway
a dummy .keep file has been added.

The alternative would have been to remove the src/ folder from
projects' classpath but since they are example projects aimed
to be use by beginners keeping the src/ folder looks like a
better solution.

* Rename "Hello World" example project's model files

Rename the files to improve consistency between the two "Hello World" projects.

* Homogenize "Hello World" projects

Make the metamodels and the semantics of those projects
more consistent to prevent confusing beginners.

* Set "*.xmi" as the default file pattern when running ALE

When ALE is run from the context menu it first opens a dialog
asking for a model instance to execute. This dialog now has
"*.xmi" for initial pattern so that XMI files are shown by
default without requiring any user typing.

An attempt was made to automatically select the XMI file that
is next to the DSL file by default (if any), but a bug seems
to prevent this feature, see:
 - https://bugs.eclipse.org/bugs/show_bug.cgi?id=214491