add ALEDynamicExpressionEvaluator
This class offers services similar to AQLSiriusInterpreter but without
the registration to the workspace listener
If offers the possibility to parse and evaluate string expressions and
return various kinds of objects
It relies on the queryEnvironment of an existing ALEInterpreter. (ie.
this is the responsibility of this ALEInterpreter
to clean/dispose the query environment when not use anymore)