Attempt to fix BehaviorLayerTest (#115)
Because:
 - broken by recent changes related to ALE environement management

Changes:
 - Sirius' Behavior layer is not automatically updated anymore and requires a manual Refresh
 - dynamic EReferences (declared in ALE) are not represented as edges anymores

Those changes are due to the fact that ALE's environment may not be stored in a .dsl configuration file anymore.
Hence, the AleResource that was used so far has been replaced by calls to an IAleEnvironment. Without the resource,
Sirius is not automatically refreshed anymore.

For the EReferences I'm not sure through. Looks like they are managed by Services::getSource and Services::getTarget
but these methods are never called and I don't know why.

Signed-off-by: Emmanuel Chebbi <emmanuel.chebbi@outlook.fr>